+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for PWA target for Angular Applications
 - Fix peer dependency checks when the dependency uses `&&` in its version range
+- Option to migrate to the Capacitor 6 beta (2)
 
 ### Version 1.63